Router and authentication set
