+++ potluck-frontend/src/components/Dashboard.js
@@ -3,15 +3,24 @@ import {connect} from 'react-redux'
 
 import PotluckCard from './PotluckCard'
 import {getPotluck, addPotluck} from '../store/actions/dashActions'
+import { useHistory } from 'react-router-dom'
 
 const Dashboard = (props)=>{
+    //axios get all items for putlucks. Prakash working on
+    //axios get all users. save to store. Micheal working on
+    const {push}= useHistory()
     useEffect(()=>{
         props.getPotluck()
-    },[props])
+    },[])
+    console.log(props)
     return(
     <div>
-        rendering
-        {/* {props.potluck.map(item=>{<PotluckCard potlucks={item}/>})} */}
+        <section>
+            {props.data.map (item=>(
+                <PotluckCard item ={item} key={item.id}/>
+            ))}
+        </section>
+        <button onClick={()=>{push('/OrganizerPage')}}>create potluck</button>
     </div>
     )
 }
@@ -19,7 +28,7 @@ const Dashboard = (props)=>{
 const mapStateToProps = state => {
     console.log(state)
     return {
-      state
+      data:state.dashboard.data
     };
   };
