+++ potluck-frontend/src/components/OrganizerPage.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const OrganizerPage = ()=>{
+    return(
+        <div>we got to organizer</div>
+    )
+}
+
+export default OrganizerPage
